+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to
 
 - ♿️(frontend) fix sidepanel accessibility aria-label #1182
 - ♿️(frontend) fix more tools heading hierarchy #1181
+- ♿️(fronted) improve button descriptions for More tools actions #1184
 
 ## [1.11.0] - 2026-03-19