+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to
 
 - 🐛(frontend) disable personal custom background while deleting #1183
 - 🐛(frontend) auto-select new custom background when not logged in #1183
+- 🐛(frontend) fix device selection not applying during conference #1156
 
 ## [1.11.0] - 2026-03-19